requirements.txt
streamlit==1.37.1
pandas==2.2.2
pyyaml==6.0.2
python-pptx==0.6.23
numpy==1.26.4
